Add some code for server<->client communication.
This is incomplete because ObjectStreamReader is incomplete and cannot read
objects as written by ObjectStreamWriter as is. I can't continue before the
reader part is more complete.

As of this commit, server actually listens for connections and client actually
sends something when it invokes a remote call. However, server does not handle
the sent data because of aforementioned incompleteness.